Mms: Add the ability to select recipients from a checkbox list
When creating a new message you have to write some letters from the
contacts you want to add and select them by clicking on it.
This can be annoying if you want to send your message to several contacts.
With this change you can manually select numbers from contacts and groups
from a checkbox list.

    * Patch Set 1: intial commit
    * Patch Set 2: total recoding, groups added
    * Patch Set 3: some code cleanup
    * Patch Set 4: UI correction
    * Patch Set 5: Some more code cleanup, reduce button padding
    * Patch Set 6: Make add button visible again

Change-Id: I94c130bd40ec5dc872446ce66c57a4eb146553af

Conflicts:
	res/drawable-hdpi/ic_recipients_add.png
	res/drawable-mdpi/ic_recipients_add.png
	res/drawable-xhdpi/ic_recipients_add.png
	res/layout/recipients_editor.xml
	res/values/cm_strings.xml
	src/com/android/mms/ui/ComposeMessageActivity.java